refactor(openmetrics): consolidate XOSTOR parsers and formatters into xostor/
